add Egg

+++ src/main/java/com/Model/Enemy.java
@@ -1,15 +1,15 @@
 package com.Model;
 
 public class Enemy extends GameObject{
-	private int HP;
+	private double HP;
 
-	public Enemy(double x, double y, double width, double height, int HP){
+	public Enemy(double x, double y, double width, double height, double HP){
 		super(x, y, width, height);
 		this.HP=HP;
 	}
 
-	public int getHP(){ return HP;}
-	public void setHP(int HP){
+	public double getHP(){ return HP;}
+	public void setHP(double HP){
 		this.HP=HP;
 		if (HP<=0) setAlive(false);
 	}
